requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-Cors==4.0.0
